coral_pitch_generator/requirements.txt
coral-sdk
mistralai
requests
elevenlabs
pillow
python-dotenv
google-generativeai